v0.99.17 new accessors, revised show(), revised taxaplot()
taxaplot() no longer S4. Should probably have an alias added to make it
plot_taxa_bar() and conform with other plot naming conventions in
phyloseq. Accessors added to make it easier to determine the names of
sample variables and taxonomic ranks, as well as the set of taxa
present at a particular taxonomic rank.