Fix igraph compatibility build error
A recent major update to the igraph package, version 0.6+, is not
backward compatible and causes certain versions of phyloseq to fail to
install properly. Although this only actually effects a small minority
of functions in phyloseq, it causes the vignettes to fail to build,
also stopping package installation.

I am still working with Bioconductor on a targeted update to the
release version of phyloseq there. I will be sure to update the devel
version on Bioconductor as well to bring that up to date with the
latest GitHub version.