Protect confusing error when tree pruned to <2 tips
Return NULL instead of "phylo" tree in both prune_species() and
merge_species(). The warning states clearly that the tree would have
been 1 or fewer tips, and is replaced with NULL. A 1-species
phyloseq-object is otherwise "allowed", though it might be useless in
practice. The warning is nevertheless important, because the
alternative error from ape::drop.tip() is unintelligible, and could
lead to substantial time waster searching for the source of the error,
if, say, you didn't realize that your merge was resulting in 1 or 0
species. Unit tests to follow.