Tree plots: ladderize, margins, color scales
This commit adds a few features to the tree plots that were requested
by a user:

1) Ladderization. This commit fixes a bug in the tree layout code, so
now the ladderize() function from ape now correctly affects the tree
layout. A ladderize parameter was added to the main plot function.

2) Margins. A plot.margin parameter was added to the plot function, to
let users add extra space to the right of the tree when long taxa
names are being truncated.

3) Color scales. A color.scale parameter was added to let users
specify their own ggplot color scale for the sample points.