add a merge_function to create the new OTU table
otu function works. now the issue is passing the objects back into a phyloseq object

passing otus in t the comp2_list works. now an error for singleton values

fixed the singleton errorwhere a vectro was being returned instead of a dataframe. now there is a discrepancy wih regard to the taxa that are being chosen.

try:
tax_glom(GP.chl, taxrank="Kingdom", NArm=FALSE)
and
tax_glom(tax_table(GP.chl), taxrank="Kingdom", NArm=FALSE)

add taxa sums to provide the correct archetype for filtering samples. but the problem here is that sometime tax_glom doens't have na OTU same so can't calculate the OTUs to use.

THe new way to do it is to merge the OTUs first and calculate the OTUS for use and pass them to the others. If there is no OTUS, just use the default