merge_samples checks for single value before coercing to numeric
Currently merge_samples forces everything to a numeric before the
(parameterized) merge function is applied to the resulting vectors.
This loses all character information (since characters are coerced to
NA) and usable factor information (since the mapping between values and
levels are lost).

This update first checks to see if there is just a single value in each
group, and if so returns that single value, before applying the current
coerce/mean approach. This preserves character values when they are the
same within the merged group.

It also gives users more power to provide a custom funciton to handle
merging, as the numeric coersion is no longer hardcoded into the
method, but is not inside the `fun` argument.