Replace sep with collapse in paste message
The `sep` argument is used to separate strings that are being pasted
together. The pasted input here is merely a vector of strings that won't
be pasted together. So the message output currently pastes them all
together like "SAMPLE1SAMPLE2SAMPLE3". This commit replaces the sep
argument with the collapse argument, which will change this behavior to
add tabs between the sample names that are removed.